move claim card to specific folder
